Click on exercise in program & fixed small bugs

+++ app/src/main/java/com/example/liftingstack/ExerciseActivities/ExerciseInstructionsPage.java
@@ -93,15 +93,15 @@ public class ExerciseInstructionsPage extends AppCompatActivity {
         if (dateDataMap != null){
             for (Map<String, Map<String, ArrayList<String>>> dateMap : dateDataMap.values()) {
                 for (Map<String, ArrayList<String>> sessionMap : dateMap.values()) {
-                    float reps = 0;
+                    int reps = 0;
                     float weight = 0;
-                    float sets = 0;
+                    int sets = 0;
                     for (ArrayList<String> setMap: sessionMap.values()){
                         sets++;
                         reps += Integer.parseInt(setMap.get(0));
-                        weight += Integer.parseInt(setMap.get(1));
+                        weight += Float.parseFloat(setMap.get(1));
                     }
-                    float aveRep = reps / sets;
+                    float aveRep = (float) reps / sets;
                     float aveWeight = (weight / sets) * 2;
                     yValues.add(new Entry(counter, aveRep*aveWeight));
                     counter++;
